issue #1 (feat- models) : move postgres_topic_dao in postgres_impl folder
